+++ modules/afsocket/afinet-source.c
@@ -26,7 +26,7 @@
 #include "misc.h"
 #include "transport-mapper-inet.h"
 #include "socket-options-inet.h"
-#include "transport/tlstransport.h"
+#include "transport/transport-tls.h"
 
 #include <sys/types.h>
 #include <sys/socket.h>
